+++ relay-server/src/endpoints/common.rs
@@ -1,13 +1,13 @@
 //! Common facilities for ingesting events through store-like endpoints.
 use std::io;
 
+use axum::extract::rejection::{BytesRejection, FailedToBufferBody};
 use axum::http::{StatusCode, header};
 use axum::response::IntoResponse;
 use bytes::Bytes;
 use chrono::Utc;
 use futures::TryStreamExt;
 use futures::stream::BoxStream;
-use multer::Field;
 use relay_config::{Config, RelayMode};
 use relay_event_schema::protocol::{EventId, EventType};
 use relay_quotas::{RateLimits, Scoping};
@@ -119,6 +119,20 @@ pub enum BadStoreRequest {
 
     #[error("project not available")]
     ProjectUnavailable,
+
+    #[error("failed to upload to objectstore")]
+    ObjectstoreUploadFailed,
+}
+
+impl From<BytesRejection> for BadStoreRequest {
+    fn from(value: BytesRejection) -> Self {
+        match value {
+            BytesRejection::FailedToBufferBody(FailedToBufferBody::LengthLimitError(_)) => {
+                BadStoreRequest::ContentTooLarge
+            }
+            other => BadStoreRequest::InvalidBody(io::Error::other(other)),
+        }
+    }
 }
 
 impl From<multer::Error> for BadStoreRequest {
@@ -174,6 +188,9 @@ impl IntoResponse for BadStoreRequest {
             }
             BadStoreRequest::Overflow(_) | BadStoreRequest::ContentTooLarge => {
                 (StatusCode::PAYLOAD_TOO_LARGE, body).into_response()
+            }
+            BadStoreRequest::ObjectstoreUploadFailed => {
+                (StatusCode::INTERNAL_SERVER_ERROR, body).into_response()
             }
             _ => {
                 // In all other cases, we indicate a generic bad request to the client and render
@@ -483,16 +500,20 @@ fn emit_envelope_metrics(envelope: &Envelope) {
 /// [AttachmentPlaceholder] as payload.
 ///
 /// Returns `None` if uploading fails due to any reason.
-pub async fn upload_to_objectstore(
-    field: Field<'static>,
+pub async fn upload_to_objectstore<S, E>(
+    stream: S,
+    content_type: Option<String>,
     mut item: Item,
     config: &Config,
     scoping: Scoping,
     upload: &Addr<Upload>,
     referrer: &'static str,
-) -> Option<Item> {
-    let content_type = field.content_type().map(ToString::to_string);
-    let stream: BoxStream<'static, io::Result<Bytes>> = Box::pin(field.map_err(io::Error::other));
+) -> Option<Item>
+where
+    S: futures::Stream<Item = Result<Bytes, E>> + Send + 'static,
+    E: Into<Box<dyn std::error::Error + Send + Sync>> + 'static,
+{
+    let stream: BoxStream<'static, io::Result<Bytes>> = Box::pin(stream.map_err(io::Error::other));
     let stream = MeteredStream::new(stream, referrer);
     let stream = BoundedStream::new(stream, 1, config.max_upload_size());
     let byte_counter = stream.byte_counter();