+++ relay-server/src/endpoints/playstation.rs
@@ -88,8 +88,10 @@ impl<'a> AttachmentStrategy for PlaystationAttachmentStrategy<'a> {
     ) -> Result<Option<Item>, multer::Error> {
         match self.upload_service {
             Some(upload) if self.infer_type(&field) != AttachmentType::Prosperodump => {
+                let content_type = field.content_type().map(ToString::to_string);
                 Ok(common::upload_to_objectstore(
                     field,
+                    content_type,
                     item,
                     config,
                     self.scoping,