+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 **Internal**:
 
-- Stream attachments received at `/minidump` to objectstore. ([#5877](https://github.com/getsentry/relay/pull/5877))
+- Stream attachments and minidumps received at `/minidump` to objectstore. ([#5877](https://github.com/getsentry/relay/pull/5877), [#5909](https://github.com/getsentry/relay/pull/5909))
 - Distinguish between objectstore stream timeouts and regular timeouts on a config level. ([#5878](https://github.com/getsentry/relay/pull/5878))